get_wallpaper: fix LXDE, Lubuntu support
PCManFM stores its config file in different places depending on the desktop session, which is normally "LXDE" but "Lubuntu" on Lubuntu. The latter is sort of a special case because the case of the session name and the config path do not match.

Also, recent versions of PCManFM store the wallpaper in desktop-items-0.conf; update to reflect that change.