SafeLogger: port to Python 3
makeRecord() got a new sinfo keyword arg in Python 3, which was not handled here.

Since we don't need to process any arguments after msg, we can actually just use *args and **kwargs to pass the rest in as is.

Also update the makeRecord() description to better reflect its actual behaviour.