Replace LogMethodCalls with a profiler-based logging system
The existing implementation broke in Python 3 because of its removal of "unbound methods" - i.e. functions
that require a class instance but are accessed as a class attribute. According to [1], this makes determining
the class of an instance method less than trivial when referred to from an instance attribute, since a bare
function object is returned instead.

This new logger works a bit differently - it still supports filtering by packages, but i doesn't log function
parameters because those are not isolated in Python frames. However, it is able to track the *exit* and line numbers
of functions, which IMO are also very useful.

[1]: https://stackoverflow.com/questions/3589311/get-defining-class-of-unbound-method-object-in-python-3