Integrate FloatingWindow with the Variety app
On right click or any double click, the floating window will pop up the Variety menu.

It supports dragging to move the icon when LMB is held, but that means we can't use a single left click to popup the menu (since that would make the window itself lose focus).