Fix blue lockscreen in GNOME
Previous to this commit when a wallpaper was changed,
under GNOME and Fedora the lockscreen wallpaper would
default to a solid blue.

Also moved the code for the lockscreen change right under
the GNOME section in `set_wallpaper` and make sure the
picture-option for the screen saver is set to 'zoom'.