Add 'Auto' option to icon and set depending on dark mode system pref
This PR is a draft to add a `PreferDarkModeListener` to Variety that can
have callbacks registered to it in order to notify parts of the app of
whether the user prefers dark mode or light mode at an OS level.

It is currently only being used to set the icon to Dark or Light
automatically. But it could be used to choose wallpapers of a certain
tone if the user wants that.

:warning: I have not tested this PR yet and am posting it just to gauge
interest. I will see if it works and productize it if people want this.