Fix kill command syntax in set_wallpaper script (#802)
* Fix kill command syntax in set_wallpaper script

`kill` was failing in cases when multiple `swaybg` instances were running simultaneously, because it was receiving a string of PIDs as a single argument. Removing the double quotes makes it effectively end any and all `swaybg` processes. Running processes were accumulating due to this fault, making this a chinken-and-egg problem. Solved, anyway.

* Loop on multiple swaybg processes

Refactor PID handling to support multiple processes using a loop, in order to comply to shellcheck SC2046.

* Fix syntax errors in set_wallpaper script